# openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Nexus REST API",
    "description": "Explore resources and get their data.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/v1/users/authenticate": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Authenticates the user.",
        "operationId": "Users_Authenticate",
        "parameters": [
          {
            "name": "scheme",
            "in": "query",
            "required": true,
            "description": "The authentication scheme to challenge.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "returnUrl",
            "in": "query",
            "required": true,
            "description": "The URL to return after successful authentication.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/signout": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Logs out the user.",
        "operationId": "Users_SignOut",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/refresh-token": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Refreshes the JWT token.",
        "operationId": "Users_RefreshToken",
        "requestBody": {
          "x-name": "request",
          "description": "The refresh token request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "A new pair of JWT and refresh token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/revoke-token": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Revokes a refresh token.",
        "operationId": "Users_RevokeToken",
        "requestBody": {
          "x-name": "request",
          "description": "The revoke token request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeTokenRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/accept-license": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Accepts the license of the specified catalog.",
        "operationId": "Users_AcceptLicense",
        "parameters": [
          {
            "name": "catalogId",
            "in": "query",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{userId}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Deletes a user.",
        "operationId": "Users_DeleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The identifier of the user.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{userId}/{claimId}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Puts a claim.",
        "operationId": "Users_PutClaim",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The identifier of the user.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "description": "The identifier of claim.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "claim",
          "description": "The claim to put.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NexusClaim"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Deletes a claim.",
        "operationId": "Users_DeleteClaim",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The identifier of the user.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "description": "The identifier of the claim.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TokenPair": {
        "type": "object",
        "description": "A token pair.",
        "additionalProperties": false,
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "The JWT token."
          },
          "refreshToken": {
            "type": "string",
            "description": "The refresh token."
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "description": "A refresh token request.",
        "additionalProperties": false,
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token."
          }
        }
      },
      "RevokeTokenRequest": {
        "type": "object",
        "description": "A revoke token request.",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string",
            "description": "The refresh token."
          }
        }
      },
...
      "NexusClaim": {
        "type": "object",
        "description": "Represents a claim.",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "description": "The claim type."
          },
          "value": {
            "type": "string",
            "description": "The claim value."
          }
        }
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Please enter 'Bearer {your JWT token}'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}